REGRESSION (r230480): Cannot adjust photo position on LinkedIn's profile page https://bugs.webkit.org/show_bug.cgi?id=186464 <rdar://problem/40369448>
Reviewed by Simon Fraser.

Source/WebCore:

The optimization logic for skipping image layout when we only need overflow computation should check if the image actually needs
simplified layout only. The needsSimplifiedNormalFlowLayout() flag means that the overflow information needs to be updated but
it does not mean that overflow is the only property that we need to recompute.

Test: fast/images/positioned-image-when-transform-is-present.html

* rendering/RenderImage.cpp:
(WebCore::RenderImage::layout):
* rendering/RenderObject.h:
(WebCore::RenderObject::needsSimplifiedNormalFlowLayoutOnly const):

LayoutTests:

* fast/images/positioned-image-when-transform-is-present-expected.html: Added.
* fast/images/positioned-image-when-transform-is-present.html: Added.


Identifier: 201820@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232663 268f45cc-cd09-0410-ab3c-d52691b4dbfc