JavaScriptCore: Disable 32-bit JIT on Windows https://bugs.webkit.org/show_bug.cgi?id=185989
Reviewed by Mark Lam.

.:

* Source/cmake/OptionsWin.cmake:

Source/JavaScriptCore:

Fixed the CLOOP so it can work when COMPUTED_GOTOs are not supported.

* llint/LLIntData.h:
(JSC::LLInt::getCodePtr): Used a reinterpret_cast since Opcode could be an int.
* llint/LowLevelInterpreter.cpp: Changed the definition of OFFLINE_ASM_GLOBAL_LABEL to not
have a case label because these aren't opcodes.
* runtime/Options.cpp: Made assembler related Windows conditional code also conditional
on the JIT being enabled.
(JSC::recomputeDependentOptions):

Source/WTF:

Fixed the CLOOP so it can work when COMPUTED_GOTOs are not supported.

* wtf/Platform.h:


Identifier: 201851@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232719 268f45cc-cd09-0410-ab3c-d52691b4dbfc