Link drag image is inconsistently unreadable in dark mode https://bugs.webkit.org/show_bug.cgi?id=186472
Reviewed by Timothy Hatcher.

Source/WebCore:

* platform/mac/DragImageMac.mm:
(WebCore::createDragImageForLink):
Use LocalDefaultSystemAppearance so that NSColors used inside
createDragImageForLink are interpreted correctly. This function
always follows the system appearance regardless of what the preference
is set to, because it's generating UI that isn't part of the page.

Use controlBackgroundColor to get a consistently contrasting background
for the link drag image.

Source/WebKit:

* UIProcess/Cocoa/WebViewImpl.mm:
(WebKit::WebViewImpl::useDefaultAppearance):
Make defaultAppearance accurate even if useSystemAppearance is false.
Some parts of WebKit (like the link drag image, but also context menus)
want to be able to follow the system appearance regardless of whether
the view or content has opted in.

Source/WebKitLegacy/mac:

* WebView/WebView.mm:
(-[WebView _defaultAppearance]):
Make defaultAppearance accurate even if useSystemAppearance is false.
Some parts of WebKit (like the link drag image, but also context menus)
want to be able to follow the system appearance regardless of whether
the view or content has opted in.


Identifier: 201863@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232731 268f45cc-cd09-0410-ab3c-d52691b4dbfc