CachedRawResource wastes 57K of Vector capacity https://bugs.webkit.org/show_bug.cgi?id=186703 <rdar://problem/41173515>
Reviewed by Simon Fraser.

Use 0 as minimum capacity for CachedRawResource::m_redirectChain, instead of the
default value of 16. m_redirectChain usually contains very few items so 16 is
wasteful, especially considering that m_redirectChain contains fairly large
objects (std::pair<ResourceRequest, ResourceResponse>). CachedRawResource are
stored in the memory cache so it is important to not use more memory than
necessary.

We are also not worried about the performance cost of allocations here since
this vector is only populated on HTTP redirects.

* loader/cache/CachedRawResource.h:


Identifier: 202001@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232897 268f45cc-cd09-0410-ab3c-d52691b4dbfc