[CMake] Automatically disable JIT and enable USE_SYSTEM_MALLOC on unfamiliar architectures https://bugs.webkit.org/show_bug.cgi?id=186722
Reviewed by Darin Adler.

We know that the JSC JIT and bmalloc both work on only a limited set of architectures. In
Fedora, we have to manually disable these when building for s390x, ppc64, and ppc64le. But
it's really easy to do the right thing automatically, so we might as well.

* Source/cmake/WebKitFeatures.cmake:


Identifier: 202021@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@232918 268f45cc-cd09-0410-ab3c-d52691b4dbfc