[WSL] Add details to the sphinx outline https://bugs.webkit.org/show_bug.cgi?id=186310
Rubberstamped by Filip Pizlo.

Mostly improving the sphinx outline.
Also started moving the execution rules to WSL_type.ott (which will eventually become WSL.ott).
Tried formalizing the gathering of declarations into the environment, but I think I will scratch it.
It is both trivial to do (and to explain in english) and a bit of a pain to formalize.




Identifier: 202113@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@233013 268f45cc-cd09-0410-ab3c-d52691b4dbfc